Change the logic how to loop across contrasts/images.
Now we iterate directly through the contrasts specified by the '-c' arg. There is thus no need to pass paths to the images from the python script.

I decided for this because the paths from the python script were pointing to the images under 'bids'. But we actually want to do the analysis on copied images to 'data_processed'. We can thus construct paths to images under 'data_processed' directly in the bash script.